Extract create-issue tests to satisfy 1000-line source file limit
The PR's test additions pushed two files over the hard 1000-line CI gate.
Move create-issue parser tests to github/tests/create_issue.rs and
create-issue reducer tests to state/issues_tests_create.rs.
