+++ src/github/tests/mod.rs
@@ -1,3 +1,4 @@
 mod actions;
 mod auth_device;
+mod create_issue;
 mod issues;
